+++ config/environments/development.js
@@ -5,5 +5,6 @@ module.exports.redis            = {host: '127.0.0.1',
                                    port: 6379, 
                                    idleTimeoutMillis: 1,
                                    reapIntervalMillis: 1};
+module.exports.mapnik_version   = undefined; // will be looked up at runtime if undefined
 module.exports.windshaft_port   = 8080;
 module.exports.enable_cors      = true;