+++ NEWS.md
@@ -1,9 +1,10 @@
 Version 0.5.9-dev
-2012-??-??
+2012-10-??
 
  - CartoCSS version control
   - Include version in GET /style response
   - Support version in POST /style request
+  - Autodetect target mapnik version and let config override it
 
 Version 0.5.8
 2012-09-28